Configure core paths with plugin instead of config
Since oh-my-zsh applies plugins before custom configurations, plugins
that use the path to find commands may fail to find the commands if
the path has not already been setup.

Workaround this issue by moving the core path setup out of custom
configuration files and into a `custom-paths` plugin. The only weird
side effect is that the plugin is added to the fpath.